+++ app/src/main/java/kr/co/bootpay/android/TotalPaymentActivity.java
@@ -43,7 +43,7 @@ public class TotalPaymentActivity extends AppCompatActivity {
         items.add(item2);
 
         Payload payload = new Payload();
-        payload.setApplicationId("5b8f6a4d396fa665fdc2b5e8")
+        payload.setApplicationId(BootpayConstants.application_id)
 //                .setPg("나이스페이")
 //                .setMethods(Arrays.asList("카드", "휴대폰"))
                 .setOrderName("부트페이 결제테스트")